moving dateFromString function of roster controller into utility function, and using for reingest date parsing. Fixing lint errors. throwing appropriate errors instead of manual error definition
