+++ server/api/reingest/index.ts
@@ -1,14 +1,13 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import controller from './reingest.controller';
-import { requireRolePermission } from '../../auth';
-
+import { requireInternalUser } from '../../auth';
 
 const router = express.Router() as any;
 
 router.post(
   '/:edipi',
-  requireRolePermission(role => role.canManageGroup),
+  requireInternalUser,
   bodyParser.json(),
   controller.reingestSymptomRecords,
 );
